fix(payload-plugin-comments): split types into separate domain files
