+++ packages/payload-plugin-comments/src/types/base.ts
@@ -1,0 +1,11 @@
+import type { Payload } from "payload";
+
+export interface BaseServiceOptions {
+  payload?: Payload;
+  locale?: string | null;
+}
+
+export interface BaseDocument {
+  id: string | number;
+  [key: string]: unknown;
+}
