+++ packages/payload-plugin-comments/src/types/collection.ts
@@ -1,0 +1,1 @@
+export type DocumentTitles = Record<string, Record<string, string>>;
