+++ packages/payload-plugin-comments/src/types/index.ts
@@ -1,0 +1,7 @@
+export type { BaseServiceOptions, BaseDocument } from "./base";
+export type { CommentMention, Comment } from "./comment";
+export type { CollectionEntry, TenantPluginConfig, CommentsPluginConfig, CommentsPluginConfigStorage } from "./config";
+export type { User } from "./user";
+export type { FilterMode, LoadingStatus, Response } from "./general";
+export type { DocumentTitles } from "./collection";
+export type { EntityConfig, EntityLabel, EntityLabelsMap, FieldLabelSegment, GlobalFieldLabelRegistry, Mode } from "./entity";
