+++ packages/payload-plugin-comments/src/types/user.ts
@@ -1,0 +1,5 @@
+export interface User {
+  id: number;
+  email?: string | null;
+  [key: string]: unknown;
+}
